+++ front/pages/api/w/[wId]/data_sources/[name]/tables/csv.ts
@@ -1,18 +1,13 @@
-import type { CoreAPIRow } from "@dust-tt/types";
-import type { Result } from "@dust-tt/types";
-import type { APIError } from "@dust-tt/types";
-import { CoreAPI } from "@dust-tt/types";
-import { Err, Ok } from "@dust-tt/types";
-import { parse } from "csv-parse";
+import { assertNever } from "@dust-tt/types";
 import { isLeft } from "fp-ts/lib/Either";
 import * as t from "io-ts";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { getDataSource } from "@app/lib/api/data_sources";
+import { upsertTableFromCsv } from "@app/lib/api/tables";
 import { Authenticator, getSession } from "@app/lib/auth";
 import { generateModelSId } from "@app/lib/utils";
-import logger from "@app/logger/logger";
 import { apiError, withLogging } from "@app/logger/withlogging";
 
 export const config = {
@@ -39,13 +34,6 @@ export type UpsertTableFromCsvRequestBody = t.TypeOf<
   typeof CreateTableFromCsvSchema
 >;
 
-export type RowValue =
-  | number
-  | boolean
-  | string
-  | { type: "datetime"; epoch: number }
-  | null;
-
 async function handler(
   req: NextApiRequest,
   res: NextApiResponse
@@ -68,11 +56,6 @@ async function handler(
     });
   }
 
-  if (!owner.flags.includes("structured_data")) {
-    res.status(404).end();
-    return;
-  }
-
   const dataSource = await getDataSource(auth, req.query.name as string);
   if (!dataSource) {
     return apiError(req, res, {
@@ -83,128 +66,11 @@ async function handler(
       },
     });
   }
+
   const coreAPI = new CoreAPI(logger);
   switch (req.method) {
     case "POST":
-      const bodyValidation = CreateTableFromCsvSchema.decode(req.body);
-      if (isLeft(bodyValidation)) {
-        const pathError = reporter.formatValidationErrors(bodyValidation.left);
-        return apiError(req, res, {
-          api_error: {
-            type: "invalid_request_error",
-            message: `Invalid request body: ${pathError}`,
-          },
-          status_code: 400,
-        });
-      }
-
-      const { name, description, csv } = bodyValidation.right;
-      const csvRowsRes = csv ? await rowsFromCsv(csv) : null;
-
-      let csvRows: CoreAPIRow[] | undefined = undefined;
-      if (csvRowsRes) {
-        if (csvRowsRes.isErr()) {
-          return apiError(req, res, {
-            api_error: csvRowsRes.error,
-            status_code: 400,
-          });
-        }
-
-        csvRows = csvRowsRes.value;
-      }
-
-      if ((csvRows?.length ?? 0) > 500_000) {
-        return apiError(req, res, {
-          api_error: {
-            type: "invalid_request_error",
-            message: `CSV has too many rows: ${csvRows?.length} (max 500_000).`,
-          },
-          status_code: 400,
-        });
-      }
-
-      const tableId = bodyValidation.right.tableId ?? generateModelSId();
-
-      const tableRes = await coreAPI.upsertTable({
-        projectId: dataSource.dustAPIProjectId,
-        dataSourceName: dataSource.name,
-        description,
-        name,
-        tableId,
-      });
-
-      if (tableRes.isErr()) {
-        logger.error(
-          {
-            dataSourceName: dataSource.name,
-            workspaceId: owner.id,
-            tableId,
-            tableName: name,
-            error: tableRes.error,
-          },
-          "Failed to upsert table."
-        );
-
-        return apiError(req, res, {
-          status_code: 500,
-          api_error: {
-            type: "internal_server_error",
-            message: "Failed to upsert table.",
-          },
-        });
-      }
-
-      if (csvRows) {
-        const rowsRes = await coreAPI.upsertTableRows({
-          projectId: dataSource.dustAPIProjectId,
-          dataSourceName: dataSource.name,
-          tableId,
-          rows: csvRows,
-          truncate: true,
-        });
-        if (rowsRes.isErr()) {
-          logger.error(
-            {
-              dataSourceName: dataSource.name,
-              workspaceId: owner.id,
-              tableId,
-              tableName: name,
-              error: rowsRes.error,
-            },
-            "Failed to upsert rows."
-          );
-
-          const delRes = await coreAPI.deleteTable({
-            projectId: dataSource.dustAPIProjectId,
-            dataSourceName: dataSource.name,
-            tableId,
-          });
-
-          if (delRes.isErr()) {
-            logger.error(
-              {
-                dataSourceName: dataSource.name,
-                workspaceId: owner.id,
-                tableId,
-                tableName: name,
-                error: delRes.error,
-              },
-              "Failed to delete table after failed upsert."
-            );
-          }
-
-          return apiError(req, res, {
-            status_code: 500,
-            api_error: {
-              type: "internal_server_error",
-              message: "Failed to upsert rows.",
-            },
-          });
-        }
-      }
-
-      return res.status(200).json({ table: tableRes.value });
-
+      return handlePostTableCsvUpsertRequest(auth, req, res);
     default:
       return apiError(req, res, {
         status_code: 405,
@@ -218,189 +84,129 @@ async function handler(
 
 export default withLogging(handler);
 
-export async function rowsFromCsv(
-  csv: string
-): Promise<Result<CoreAPIRow[], APIError>> {
-  // Detect the delimiter: try to parse the first 2 lines with different delimiters,
-  // keep the one that works for both lines and has the most columns.
-  let delimiter: string | undefined = undefined;
-  let delimiterColsCount = 0;
-  for (const d of [",", ";", "\t"]) {
-    const records: unknown[][] = [];
-    try {
-      const parser = parse(csv, { delimiter: d });
-      for await (const record of parser) {
-        records.push(record);
-        if (records.length == 2) {
-          break;
-        }
-      }
-    } catch (e) {
-      // Ignore error.
-      continue;
-    }
-
-    const [firstRecord, secondRecord] = records;
-    if (!!firstRecord.length && firstRecord.length === secondRecord.length) {
-      if (firstRecord.length > delimiterColsCount) {
-        delimiterColsCount = firstRecord.length;
-        delimiter = d;
-      }
-    }
-  }
-
-  if (!delimiter) {
-    return new Err({
-      type: "invalid_request_error",
-      message: `Could not detect delimiter.`,
+export async function handlePostTableCsvUpsertRequest(
+  auth: Authenticator,
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  const owner = auth.workspace();
+  const plan = auth.plan();
+  if (!owner || !plan || !auth.isBuilder()) {
+    return apiError(req, res, {
+      status_code: 404,
+      api_error: {
+        type: "data_source_not_found",
+        message: "The data source you requested was not found.",
+      },
     });
   }
 
-  const parser = parse(csv, { delimiter });
-  let header: string[] | undefined = undefined;
-  const valuesByCol: Record<string, string[]> = {};
-
-  for await (const anyRecord of parser) {
-    // Assert that record is string[].
-    if (!Array.isArray(anyRecord)) {
-      throw new Error("Record is not an array");
-    }
-    if (anyRecord.some((r) => typeof r !== "string")) {
-      throw new Error("Record contains non-string values");
-    }
-
-    const record = anyRecord as string[];
-
-    if (!header) {
-      header = record;
-      continue;
-    }
-
-    header = header.map((h) => h.trim().toLowerCase());
-    const headerSet = new Set<string>();
-    for (const h of header) {
-      if (headerSet.has(h)) {
-        return new Err({
-          type: "invalid_request_error",
-          message: `Duplicate header: ${h}.`,
-        });
-      }
-      headerSet.add(h);
-    }
-
-    for (const [i, h] of header.entries()) {
-      const col = record[i];
-      if (col === undefined) {
-        return new Err({
-          type: "invalid_request_error",
-          message: `Invalid record length at row ${i} (expected ${header.length} columns, got ${record.length})).`,
-        });
-      }
-      if (!valuesByCol[h]) {
-        valuesByCol[h] = [col];
-      } else {
-        (valuesByCol[h] as string[]).push(col);
-      }
-    }
+  if (!owner.flags.includes("structured_data")) {
+    res.status(404).end();
+    return;
   }
-
-  if (!header || !Object.values(valuesByCol).some((vs) => vs.length > 0)) {
-    return new Err({
-      type: "invalid_request_error",
-      message: `CSV is empty.`,
+    apiError(req, res, {
+      status_code: 404,
+      api_error: {
+        type: "data_source_not_found",
+        message: "The data source you requested was not found.",
+      },
     });
   }
 
-  // Parse values and infer types for each column.
-  const parsedValuesByCol: Record<string, RowValue[]> = {};
-  for (const [col, values] of Object.entries(valuesByCol)) {
-    if (values.every((v) => v === "")) {
-      // All values are empty, we skip this column.
-      continue;
-    }
+  const dataSourceName = req.query.name;
 
-    // We keep the parsed values from the first parser that succeeds for all non-null values in the column.
-    parsedValuesByCol[col] = (() => {
-      for (const parser of [
-        // number
-        (v: string) =>
-          /^-?\d+(\.\d+)?$/.test(v.trim()) ? parseFloat(v.trim()) : undefined,
-        // date/datetime
-        (v: string) => {
-          const date = new Date(v.trim());
-          const epoch = date.getTime();
-          return isNaN(epoch)
-            ? undefined
-            : {
-                type: "datetime" as const,
-                epoch,
-              };
-        },
-        // bool
-        (v: string) => {
-          const lowerV = v.toLowerCase();
-          return lowerV === "true"
-            ? true
-            : lowerV === "false"
-            ? false
-            : undefined;
-        },
-        // string
-        (v: string) => v,
-      ]) {
-        const parsedValues = [];
-        for (const v of values) {
-          if (v === "") {
-            parsedValues.push(null);
-            continue;
-          }
-
-          const parsedV = parser(v);
-
-          if (parsedV === undefined) {
-            // move onto next parser
-            break;
-          }
-
-          parsedValues.push(parsedV);
-        }
-
-        if (parsedValues.length === values.length) {
-          return parsedValues;
-        }
-      }
-
-      throw new Error(
-        `Unreachable: could not infer type for column ${col}. Values: ${JSON.stringify(
-          values
-        )}`
-      );
-    })();
-  }
-
-  const nbRows = (Object.values(parsedValuesByCol)[0] || []).length;
-  const rows: CoreAPIRow[] = [];
-  for (let i = 0; i < nbRows; i++) {
-    const record = header.reduce((acc, h) => {
-      const parsedValues = parsedValuesByCol[h];
-      acc[h] =
-        parsedValues && parsedValues[i] !== undefined ? parsedValues[i] : "";
-      return acc;
-    }, {} as Record<string, RowValue>);
-
-    const rowId = record["__dust_id"] ?? i.toString();
-
-    if (typeof rowId !== "string") {
-      return new Err({
+  if (!dataSourceName || typeof dataSourceName !== "string") {
+    return apiError(req, res, {
+      status_code: 400,
+      api_error: {
         type: "invalid_request_error",
-        message: `Invalid row id: ${rowId}.`,
+        message: "The data source name is missing.",
+      },
+    });
+  }
+
+  const dataSource = await getDataSource(auth, dataSourceName);
+  if (!dataSource) {
+    return apiError(req, res, {
+      status_code: 404,
+      api_error: {
+        type: "data_source_not_found",
+        message: "The data source you requested was not found.",
+      },
+    });
+  }
+
+  const bodyValidation = CreateTableFromCsvSchema.decode(req.body);
+  if (isLeft(bodyValidation)) {
+    const pathError = reporter.formatValidationErrors(bodyValidation.left);
+    return apiError(req, res, {
+      api_error: {
+        type: "invalid_request_error",
+        message: `Invalid request body: ${pathError}`,
+      },
+      status_code: 400,
+    });
+  }
+
+  const { name, description, csv } = bodyValidation.right;
+  const tableId = bodyValidation.right.tableId ?? generateModelSId();
+
+  const tableRes = await upsertTableFromCsv({
+    owner,
+    projectId: dataSource.dustAPIProjectId,
+    dataSourceName,
+    tableName: name,
+    tableDescription: description,
+    tableId,
+    csv: csv ?? null,
+  });
+
+  if (tableRes.isErr()) {
+    if (tableRes.error.type === "internal_server_error") {
+      return apiError(req, res, {
+        api_error: {
+          type: "internal_server_error",
+          message: tableRes.error.message,
+        },
+        status_code: 500,
       });
     }
 
-    delete record["__dust_id"];
+    if (tableRes.error.type === "invalid_request_error") {
+      if ("csvParsingError" in tableRes.error) {
+        return apiError(req, res, {
+          api_error: {
+            type: "invalid_request_error",
+            message: `Failed to parse CSV: ${tableRes.error.csvParsingError.message}`,
+          },
+          status_code: 400,
+        });
+      } else if ("inputValidationError" in tableRes.error) {
+        return apiError(req, res, {
+          api_error: {
+            type: "invalid_request_error",
+            message: `Invalid request body: ${tableRes.error.inputValidationError}`,
+          },
+          status_code: 400,
+        });
+      } else {
+        assertNever(tableRes.error);
+      }
+    }
 
-    rows.push({ row_id: rowId, value: record });
+    if (tableRes.error.type === "not_found_error") {
+      return apiError(req, res, {
+        api_error: {
+          type: tableRes.error.notFoundError.type,
+          message: tableRes.error.notFoundError.message,
+        },
+        status_code: 404,
+      });
+    }
+
+    assertNever(tableRes.error);
   }
 
-  return new Ok(rows);
+  return res.status(200).json(tableRes.value);
 }
